requirements.txt
django
regex
json-logic-qubit
